Integrates get_genotype in paddleslim (#228)
